refactor(types): перенести helpers (customAttr/std/owner/tabSect/mdoRef) в MCE — снизить coupling MCS до 17
