+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/registry/MetadataChildrenExtractor.java
@@ -21,8 +21,10 @@
  */
 package com.github._1c_syntax.bsl.languageserver.types.registry;
 
+import com.github._1c_syntax.bsl.context.api.KnownStandardAttributes;
 import com.github._1c_syntax.bsl.mdo.AccountingRegister;
 import com.github._1c_syntax.bsl.mdo.AccumulationRegister;
+import com.github._1c_syntax.bsl.mdo.Attribute;
 import com.github._1c_syntax.bsl.mdo.AttributeOwner;
 import com.github._1c_syntax.bsl.mdo.CalculationRegister;
 import com.github._1c_syntax.bsl.mdo.ChartOfAccounts;
@@ -33,16 +35,21 @@ import com.github._1c_syntax.bsl.mdo.Enum;
 import com.github._1c_syntax.bsl.mdo.FormOwner;
 import com.github._1c_syntax.bsl.mdo.InformationRegister;
 import com.github._1c_syntax.bsl.mdo.MD;
+import com.github._1c_syntax.bsl.mdo.TabularSection;
 import com.github._1c_syntax.bsl.mdo.TabularSectionOwner;
 import com.github._1c_syntax.bsl.mdo.Task;
 import com.github._1c_syntax.bsl.mdo.TemplateOwner;
+import com.github._1c_syntax.bsl.mdo.support.AttributeKind;
+import com.github._1c_syntax.bsl.types.MDOType;
+import com.github._1c_syntax.bsl.types.MdoReference;
+import org.jspecify.annotations.Nullable;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 
 import static com.github._1c_syntax.bsl.languageserver.types.registry.MetadataCollectionSpecializer.ChildName;
-import static com.github._1c_syntax.bsl.languageserver.types.registry.MetadataCollectionSpecializer.mdoReferenceNames;
-import static com.github._1c_syntax.bsl.languageserver.types.registry.MetadataCollectionSpecializer.singleLingualMdNames;
-import static com.github._1c_syntax.bsl.languageserver.types.registry.MetadataCollectionSpecializer.tabularSectionEntries;
 
 /**
  * Извлечение имён детей конкретной коллекции из MDO-объекта. Вынесено из
@@ -51,12 +58,127 @@ import static com.github._1c_syntax.bsl.languageserver.types.registry.MetadataCo
  */
 final class MetadataChildrenExtractor {
 
+  private static final Map<MDOType, String> OWNER_TYPE_BY_MDO_TYPE = Map.ofEntries(
+    Map.entry(MDOType.CATALOG, "ОбъектМетаданных: Справочник"),
+    Map.entry(MDOType.DOCUMENT, "ОбъектМетаданных: Документ"),
+    Map.entry(MDOType.CHART_OF_ACCOUNTS, "ОбъектМетаданных: ПланСчетов"),
+    Map.entry(MDOType.CHART_OF_CALCULATION_TYPES, "ОбъектМетаданных: ПланВидовРасчета"),
+    Map.entry(MDOType.CHART_OF_CHARACTERISTIC_TYPES, "ОбъектМетаданных: ПланВидовХарактеристик"),
+    Map.entry(MDOType.EXCHANGE_PLAN, "ОбъектМетаданных: ПланОбмена"),
+    Map.entry(MDOType.TASK, "ОбъектМетаданных: Задача"),
+    Map.entry(MDOType.BUSINESS_PROCESS, "ОбъектМетаданных: БизнесПроцесс"),
+    Map.entry(MDOType.ENUM, "ОбъектМетаданных: Перечисление"),
+    Map.entry(MDOType.INFORMATION_REGISTER, "ОбъектМетаданных: РегистрСведений"),
+    Map.entry(MDOType.ACCUMULATION_REGISTER, "ОбъектМетаданных: РегистрНакопления"),
+    Map.entry(MDOType.ACCOUNTING_REGISTER, "ОбъектМетаданных: РегистрБухгалтерии"),
+    Map.entry(MDOType.CALCULATION_REGISTER, "ОбъектМетаданных: РегистрРасчета"),
+    Map.entry(MDOType.TABULAR_SECTION, "ОбъектМетаданных: ТабличнаяЧасть")
+  );
+
   private MetadataChildrenExtractor() {
+  }
+
+  static List<ChildName> singleLingualMdNames(Collection<? extends MD> items) {
+    var result = new ArrayList<ChildName>(items.size());
+    for (var item : items) {
+      var entry = ChildName.of(item.getName());
+      if (entry != null) {
+        result.add(entry);
+      }
+    }
+    return List.copyOf(result);
+  }
+
+  static List<ChildName> customAttributeNames(Collection<? extends Attribute> items) {
+    var result = new ArrayList<ChildName>(items.size());
+    for (var item : items) {
+      if (item.getKind() == AttributeKind.STANDARD) {
+        continue;
+      }
+      var entry = ChildName.of(item.getName());
+      if (entry != null) {
+        result.add(entry);
+      }
+    }
+    return List.copyOf(result);
+  }
+
+  static List<ChildName> standardAttributesFor(MD md) {
+    var ownerType = ownerTypeFor(md);
+    if (ownerType == null) {
+      return List.of();
+    }
+    var names = KnownStandardAttributes.forOwner(ownerType);
+    if (names.isEmpty()) {
+      return List.of();
+    }
+    var result = new ArrayList<ChildName>(names.size());
+    for (var n : names) {
+      var entry = ChildName.bilingual(n.getName(), n.getAlias());
+      if (entry != null) {
+        result.add(entry);
+      }
+    }
+    return List.copyOf(result);
+  }
+
+  static @Nullable String ownerTypeFor(MD md) {
+    var byMdoType = OWNER_TYPE_BY_MDO_TYPE.get(md.getMdoType());
+    if (byMdoType != null) {
+      return byMdoType;
+    }
+    if (md instanceof TabularSection) {
+      return OWNER_TYPE_BY_MDO_TYPE.get(MDOType.TABULAR_SECTION);
+    }
+    return null;
+  }
+
+  static boolean hasStandardAttributes(MD md) {
+    return ownerTypeFor(md) != null;
+  }
+
+  static List<ChildName> tabularSectionEntries(Collection<? extends TabularSection> sections) {
+    var result = new ArrayList<ChildName>(sections.size());
+    for (var ts : sections) {
+      var entry = ChildName.of(ts.getName(), ts);
+      if (entry != null) {
+        result.add(entry);
+      }
+    }
+    return List.copyOf(result);
+  }
+
+  static List<ChildName> mdoReferenceNames(Collection<MdoReference> refs) {
+    var result = new ArrayList<ChildName>(refs.size());
+    for (var ref : refs) {
+      var entry = mdoReferenceChildName(ref);
+      if (entry != null) {
+        result.add(entry);
+      }
+    }
+    return List.copyOf(result);
+  }
+
+  private static @Nullable ChildName mdoReferenceChildName(MdoReference ref) {
+    var qualifiedName = ref.getMdoRefRu();
+    if (qualifiedName.isBlank()) {
+      qualifiedName = ref.getMdoRef();
+    }
+    if (qualifiedName.isBlank()) {
+      return null;
+    }
+    var dot = qualifiedName.lastIndexOf('.');
+    var bare = dot < 0 ? qualifiedName : qualifiedName.substring(dot + 1);
+    var mdoTypeRu = ref.getType().fullName().getRu();
+    if (!mdoTypeRu.isBlank()) {
+      return ChildName.withReturnType(bare, "ОбъектМетаданных: " + mdoTypeRu + "." + bare);
+    }
+    return ChildName.of(bare);
   }
 
   static List<ChildName> attributesFor(MD md) {
     return md instanceof AttributeOwner ao
-      ? MetadataCollectionSpecializer.customAttributeNames(ao.getAllAttributes())
+      ? customAttributeNames(ao.getAllAttributes())
       : List.of();
   }
